sam-lambda/requirements.txt
boto3>=1.34.0
line-bot-sdk==3.14.2